- if you are configuring something to be executed after bunch of JobSteps, you need to give Transition decider. in our case, we want the endStep to be run even for partial success. - check out my answer here : https://stackoverflow.com/questions/37192240/spring-batch-flowjobbuilder-parallel-and-sequential-execution-of-jobstep-typ/58529755#58529755
I am going to primarily focus on two things
- within finalStep(), how can I get the output of other jobs at the CURRENT job execution context. so that when I am retrying, I will know exactly which etl I have to work on. you should also consider the fact that the final step itself could have failed. in that case, you need to be able to store the cursor in execution context so for any re-run you should look at cursor of previous context and also proceed from there. phew!
- DO NOT expose the individual etls to outside world. They can potential run the child job with a job parameter that is not yet picked by the data-sync, and when the data-sync actually runs, that run would fail. duh!
also these individual etls are useless unless the data is actually sent.
So you need Pricing-etl+finalStep. That would look if pricing-etl is already done, if so it will pick up the file name from that past run and send the data to trinity. If the file is not present then it will send an error and then the use should run it in override mode.